remove test runner workaround, fix TS config and reorganise test deps for direct compat with Tape
