--- tests/src/index.ts
@@ -1,14 +1,0 @@
-
-import neighbourhood from './sensemaker_dna/sensemaker/neighbourhood';
-import sensemaker_community_activator from './sensemaker_dna/sensemaker/community_activator';
-import sensemaker_config from './sensemaker_dna/sensemaker/sm_config';
-import applet_config from './sensemaker_dna/sensemaker/applet_config';
-import range from './sensemaker_dna/sensemaker/range';
-import dashboard from './sensemaker_dna/sensemaker/dashboard';
-range();
-neighbourhood();
-applet_config();
-sensemaker_config();
-sensemaker_community_activator();
-dashboard();
-
