+++ tests/src/utils.ts
@@ -3,12 +3,15 @@ import {
   AppInfo,
   InstallAppRequest,
   encodeHashToBase64,
+  CellId,
   CellInfo,
   ProvisionedCell,
   CellType
 } from "@holochain/client";
 import {
   Conductor,
+  createConductor,
+  addAllAgentsToAllConductors,
 } from "@holochain/tryorama";
 import { AppletConfigInput } from "@neighbourhoods/client";
 import path from "path";
@@ -25,6 +28,51 @@ export const testProviderDna = path.join(
   __dirname,
   "../../dnas/test_provider/workdir/test_provider_dna.dna"
 );
+
+export const setUpAliceandBob = async (
+  with_config: boolean = false,
+  resource_base_type?: any
+) => {
+  const alice = await createConductor();
+  const bob = await createConductor();
+  const {
+    agentsHapps: alice_happs,
+    agent_key: alice_agent_key,
+    ss_cell_id: ss_cell_id_alice,
+    provider_cell_id: provider_cell_id_alice,
+  } = await installAgent(
+    alice,
+    "alice",
+    undefined,
+    with_config,
+    resource_base_type
+  );
+  const {
+    agentsHapps: bob_happs,
+    agent_key: bob_agent_key,
+    ss_cell_id: ss_cell_id_bob,
+    provider_cell_id: provider_cell_id_bob,
+  } = await installAgent(
+    bob,
+    "bob",
+    alice_agent_key,
+    with_config,
+    resource_base_type
+  );
+  await addAllAgentsToAllConductors([alice, bob]);
+  return {
+    alice,
+    bob,
+    alice_happs,
+    bob_happs,
+    alice_agent_key: alice_agent_key as Uint8Array,
+    bob_agent_key: bob_agent_key as Uint8Array,
+    ss_cell_id_alice: ss_cell_id_alice as CellId,
+    ss_cell_id_bob: ss_cell_id_bob as CellId,
+    provider_cell_id_alice: provider_cell_id_alice as CellId,
+    provider_cell_id_bob: provider_cell_id_bob as CellId,
+  };
+};
 
 export const installAgent = async (
   conductor: Conductor,
