+++ tests/tsconfig.json
@@ -3,15 +3,26 @@
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
     /* Basic Options */
     "outDir": "dist",
-    "target": "ES2017",
-    "module": "ESNext",
+    "target": "ES2018",
+    "module": "esNext",
     "moduleResolution": "node",
+    "lib": ["es2017", "dom"],
+    "sourceMap": true,
     /* Strict Type-Checking Options */
     "strict": true /* Enable all strict type-checking options. */,
     "esModuleInterop": true /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */,
+    "allowSyntheticDefaultImports": true,
     "noImplicitAny": false /* Raise error on expressions and declarations with an implied 'any' type. */,
     /* Advanced Options */
     "skipLibCheck": true /* Skip type checking of declaration files. */,
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
-  }
+  },
+  "include": [
+    "src/**/*.ts",
+    "../client/src/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "../node_modules"
+  ]
 }
